% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{breeding_wiwa_isotopes}
\alias{breeding_wiwa_isotopes}
\title{Isotope values, latitude, longitude and more data from 357 breeding Wilson's warblers}
\format{A data frame with 357 rows and 14 variables. The relevant variables for
analyses here are:
\describe{
  \item{Field_Num}{unique identifier for each bird}
  \item{Isotope.Value}{hydrogen isotope ratios measured in the bird's feather}
  \item{Lat}{latitude of the bird's breeding/sampling location}
  \item{Long}{latitude of the bird's breeding/sampling location}
}}
\source{
Kristen Ruegg, Jeff Kelly, Thomas Smith
}
\usage{
breeding_wiwa_isotopes
}
\description{
A data frame containing hydrogen isotope values, lat, long, and IDs and some other
columns of data for birds sampled on the breeding grounds.
}
\keyword{datasets}

