% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine-types.R
\name{comboize}
\alias{comboize}
\title{combine genetics, isotopes, and habitat raster with exponents as given}
\usage{
comboize(Mgen, Miso, Mhab, beta_gen, beta_iso, beta_hab)
}
\arguments{
\item{Mgen}{the genetic posteriors rasterStack.  Must be a rasterStack}

\item{Miso}{the isotope posteriors rasterStack.}

\item{Mhab}{a single layer raster with the habitat suitabiilty measure as a normalized
probability surface.}

\item{beta_gen}{the exponent to raise the habitat raster to}

\item{beta_iso}{the exponent to raise the isotope raster to}

\item{beta_hab}{the exponent to raise the genetic raster to}
}
\description{
This just multiplies the rasters together, each raised to the appropriate
exponent, normalizes and returns the result
}

