% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_isotope_posteriors}
\alias{example_isotope_posteriors}
\title{Output of \code{\link{isotope_posterior_probs}} for two migrant birds.}
\format{
An object of class \code{list} of length 2.
}
\source{
Ruegg et al 2014
}
\usage{
example_isotope_posteriors
}
\description{
Because it takes too long to generate this output for future examples, we
just store it as a data object to use in examples.  See the example in
\code{\link{isotope_posterior_probs}} to see what this is.
}
\keyword{datasets}
