% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaiah.R
\docType{package}
\name{gaiah}
\alias{gaiah}
\alias{gaiah-package}
\title{gaiah: Genetic And Isotopic Assignment accounting for Habitat suitability}
\description{
Tools for using genetic markers, stable isotope data, and habtitat
suitability data to calculate posterior probabilities of breeding origin of
migrating birds.
}
\section{Section one}{

Put something here
}

