% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genetic_regions}
\alias{genetic_regions}
\title{RasterStack showing the 6 genetic regions that Wilson's warblers may be assigned to}
\format{
RasterStack with 6 layers. Each contains 1's in the genetic region and 0's elsewhere.
The sum of these layers is the raster \code{\link{wiwa_breed}}.
\describe{
\item{class}{ RasterStack }
\item{dimensions}{ 80, 228, 18240, 6  (nrow, ncol, ncell, nlayers)}
\item{resolution}{ 0.5, 0.5  (x, y)}
\item{extent}{ -168.1, -54.1, 31.2, 71.2  (xmin, xmax, ymin, ymax)}
\item{coord. ref.}{ +proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs +towgs84=0,0,0 }
\item{data source}{in memory}
\item{names}{ CalSierra, Basin.Rockies, Eastern, AK.EastBC.AB, Wa.To.NorCalCoast, CentCalCoast }
}
}
\source{
Ruegg et al 2014
}
\usage{
genetic_regions
}
\description{
The sum over layers gives the same as \code{\link{wiwa_breed}}
}
\keyword{datasets}
