% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_wrld_simpl}
\alias{get_wrld_simpl}
\title{return the wrld_simpl data set from maptools}
\usage{
get_wrld_simpl()
}
\description{
I define this as a function so that we don't have to attach
maptools, but we can just have it in the imports. Couldn't figure
out how to do it otherwise.
}
\examples{
ws <- get_wrld_simpl()
head(ws)
\dontrun{plot(ws)}
}
