% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{hpd_cumul}
\alias{hpd_cumul}
\title{finds the cumulative posterior probability of each cell for HPD calculation}
\usage{
hpd_cumul(B)
}
\arguments{
\item{B}{a raster of posterior probs}
}
\description{
This function is not exported currently, but is used in min_hpd_inclusion_area
}
\keyword{internal}
