% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isomap_job54152_prediction}
\alias{isomap_job54152_prediction}
\title{Predicted isotope values from ISOMAP}
\format{
A tibble with 10,786 rows and 12 variables. The relevant variables for
analyses here are:
\describe{
  \item{lat}{latitude of the predicted location}
  \item{long}{longitude of the predicted location}
  \item{predreg}{Fill in}
  \item{stdreg}{Fill in}
  \item{predkrig}{Fill in}
  \item{stdkrig}{Fill in}
}
}
\source{
Kristina Paxton and ISOMAP (http://isomap.rcac.purdue.edu:8080/gridsphere/gridsphere)
}
\usage{
isomap_job54152_prediction
}
\description{
A data frame containing predicted hydrogen isotope values, lat, long, and IDs and some other
columns of data prections made by ISOMAP
}
\keyword{datasets}
