% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope-high-level.R
\name{isotope_posterior_probs}
\alias{isotope_posterior_probs}
\title{compute posterior probabilities of origin given isotope values}
\usage{
isotope_posterior_probs(isoscape, ref_birds, assign_birds = NULL,
  isoscape_pred_column = "predkrig", isoscape_sd_column = "stdkrig",
  self_assign = FALSE)
}
\arguments{
\item{isoscape}{the data frame read in from "prediction.txt" from ISOMAP. The
latitude column must be named "lat" and the longitude column must be named "long".  You have to choose
which columns to use with the parameters \code{isoscape_pred_column} and \code{isoscape_sd_column}.}

\item{ref_birds}{a data frame of reference birds. This should have (at least) columns of "ID"
(for unique identifiers for each bird), "lat", "long",
"Isotope.Value" and "Location".  The "Location" column will be used to group samples for
the Vander Zanden Rescaling.}

\item{assign_birds}{A data frame of birds whose breeding origins are to be inferred.  These must have
at a minimum the column "ID" (for uniqe identifiers for the birds) and the column "Isotope.Value". This can
be left NULL if there are no birds of unknown origin to assign (for example if you are performing cross-validation
on the ref_birds).}

\item{isoscape_pred_column}{the name of the column in \code{isoscape} to be used as the prediction (default
is "predkrig").}

\item{isoscape_sd_column}{the name of the column in \code{isoscape} to be used as the standard deviation (default
is "stdkrig").}

\item{self_assign}{if TRUE, then the birds in \code{ref_birds} will each have posterior surfaces computed for
them using a leave one out procedure (i.e. each bird in turn is left out while rescaling the precip isomap to
a tissue isomap).  Should not be TRUE if assign_birds is non NULL.}
}
\description{
This takes care of the whole shebang.  More description later
}

