% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{migrant_wiwa_genetic_posteriors}
\alias{migrant_wiwa_genetic_posteriors}
\title{Posterior probs of genetic region of origin for 926 WIWAs sampled during migration}
\format{A tbl_df-ed (from dplyr) data frame with 357 rows and 14 variables. The relevant variables for
analyses here are:
\describe{
  \item{ID}{unique identifier for each bird}
  \item{Short_Name}{same id for the bird}
  \item{Collection_Date}{The date the bird was sampled.}
  \item{NumberOfLoci}{Number of loci successfully typed}
  \item{region}{one of the genetic regions}
  \item{posterior}{the posterior prob of originating from that region}
}}
\source{
Kristina Paxton, Kristen Ruegg, Eric Anderson, Thomas Smith
}
\usage{
migrant_wiwa_genetic_posteriors
}
\description{
A long
format data frame with 5,556 rows and 6 columns
}
\keyword{datasets}

