% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{migrant_wiwa_isotopes}
\alias{migrant_wiwa_isotopes}
\title{Isotope values, latitude, longitude and more data from 357 breeding Wilson's warblers}
\format{A tbl_df-ed (from dplyr) data frame with 688 rows and 14 variables. The relevant variables for
analyses here are:
\describe{
  \item{ID}{unique identifier for each bird}
  \item{Isotope.Value}{hydrogen isotope ratios measured in the bird's feather}
}}
\source{
Kristina Paxton
}
\usage{
migrant_wiwa_isotopes
}
\description{
A data frame containing hydrogen isotope values, lat, long, and IDs and some other
columns of data for birds sampled during migration from Arizona.
}
\keyword{datasets}

