% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{min_hpd_inclusion_area}
\alias{min_hpd_inclusion_area}
\title{compute the smallest area of an HPD credible set that includes the true bird location}
\usage{
min_hpd_inclusion_area(M, lat, long)
}
\arguments{
\item{M}{the posterior surface matrix}

\item{lat}{the true latitude}

\item{long}{the true longitude}
}
\description{
This function computes the area of the HPD credible set that is as small as possible
but just barely includes the cell with the true location (or, if the true location is not
actually inluded in the range map, we use the closest cell on the range map to the true
location).  This operates on a single rasterLayer.  To operate on a rasterBrick of posteriors
and a data frame of lat-longs, use `min_hpd_inc_area_df()`.
}
\keyword{internal}
