% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope-high-level.R
\name{private_rescale}
\alias{private_rescale}
\title{internal function for isotope machinations}
\usage{
private_rescale(birds, pred, std)
}
\arguments{
\item{birds}{a data frame like "ref_birds" in \code{isotope_posterior_probs} after the isopredictions
have been attached to it.}

\item{pred}{The raster of isoscape predictions}

\item{std}{The raster of isoscape standard deviations}
}
\description{
does these things:
1. removes birds in Locations that have only 1 bird in them (printing a warning message while it does so)
2. does the vza rescaling process and returns the output that is needed for the vza_assign
   function.  (basically the ouput of the mean and var rasters). Should return a-bar and b-bar
   too --- all of that in a big list.
Then with that output I can pump it all into vza_assign for a particular (left-out) ref_bird or for
an assign bird.  I should return the fitted model for each one.
}
\keyword{internal}
