% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vander-zanden-appendix.R
\name{vza_assign}
\alias{vza_assign}
\title{assign posterior probability of origin for a bird in each cell in the raster}
\usage{
vza_assign(rescale_mean, rescale_var, precip_sd, sd_indiv, bird_iso)
}
\arguments{
\item{rescale_mean}{the tissue specific mean raster, such as the mean.raster
component of the output of \code{vza_mean_and_var_rasters}.}

\item{rescale_var}{tissue specific raster of standard deviations, such as the var.raster
component of the output of \code{vza_mean_and_var_rasters}.}

\item{precip_sd}{SD raster associated with the IsoMAP output.
This is the precip component of the variance term.}

\item{sd_indiv}{the individual component of the variance term.
This is a value to be determined by the user.  The standard approach is to
use the mean of the SDs observed among individuals at all of the calibration sites.}

\item{bird_iso}{a single value giving the isotope ratio found in the individual's feather.}
}
\description{
This is a rewrite of the function \code{assignment} from the Vander
Zanden appendix code.
}
\details{
This is a fairly low-level function.  It returns a raster of posterior probs (they are
scaled to sum to one over all cells).
}
\keyword{internal}
