% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vander-zanden-appendix.R
\name{vza_mean_and_sd_rasters}
\alias{vza_mean_and_sd_rasters}
\title{calculate a raster of mean and SD of expected tissue isotopes from precip data and resampled regressions}
\usage{
vza_mean_and_sd_rasters(iso_raster, si)
}
\arguments{
\item{iso_raster}{the raster of isotope precipitation values, for example, like that
produced by \code{\link{isomap2raster}}.}

\item{si}{slopes and intercepts from the resampled regressions.  This is a data frame
with columns named "slopes" and "intercepts" like that returned by \code{\link{vza_rescale}}}
}
\description{
This is a rewrite of the function \code{raster.conversion} from the Vander Zanden
appendix.
}

