% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiwa_breed}
\alias{wiwa_breed}
\title{a raster of the breeding range of Wilson's warbler}
\format{
This a rasterized version of the breeding range of Wilson's warbler
It contains 1's in the breeding range and 0's elsewhere.
\describe{
\item{class}{RasterLayer}
\item{dimensions}{80, 228, 18240  (nrow, ncol, ncell)}
\item{resolution}{0.5, 0.5  (x, y)}
\item{extent}{-168.1, -54.1, 31.2, 71.2  (xmin, xmax, ymin, ymax)}
\item{coord. ref.}{+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs +towgs84=0,0,0}
\item{data source}{in memory}
\item{names}{layer}
\item{values}{0, 1  (min, max)}
}
}
\source{
The rasters were generated from shapefiles provided to us by
BirdLife International. (BirdLife International and NatureServe (2012)
Bird species distribution maps of the world. BirdLife International, Cambridge,
UK and NatureServe, Arlington, USA). Persons interested in the range map
should contact BirdLife International http://www.birdlife.org/ or
NatureServe http://www.natureserve.org/ directly.
}
\usage{
wiwa_breed
}
\description{
a raster of the breeding range of Wilson's warbler
}
\keyword{datasets}
