% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiwa_habitat_unclipped}
\alias{wiwa_habitat_unclipped}
\title{RasterLayer showing the MaxEnt habitat suitability model unclipped by the known breeding range}
\format{
An object of class \code{RasterLayer} of dimension 80 x 228 x 1.
}
\source{
Ryan Harrigan
}
\usage{
wiwa_habitat_unclipped
}
\description{
\describe{
\item{class}{ RasterLayer }
\item{dimensions}{ 80, 228, 18240, 6  (nrow, ncol, ncell, nlayers)}
\item{resolution}{ 0.5, 0.5  (x, y)}
\item{extent}{ -168.1, -54.1, 31.2, 71.2  (xmin, xmax, ymin, ymax)}
\item{coord. ref.}{ +proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs +towgs84=0,0,0 }
\item{data source}{in memory}
\item{values}{0, 0.001093349  (min, max)}
}
}
\keyword{datasets}
