% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wrld_simpl}
\alias{wrld_simpl}
\title{Simple map of the world}
\format{
a SpatialPolygonsDataFrame
}
\source{
Got this from the old maptools package.  See ?maptools::wrld_simpl
}
\usage{
wrld_simpl
}
\description{
This is wrld_simpl from the maptools package.  It was all that I used from
the maptools package which is going to be archived at the end of 2023.  So,
I just saved wrld_simpl as a data object in this package.
}
\keyword{datasets}
