% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{breeding_wiwa_isotopes}
\alias{breeding_wiwa_isotopes}
\title{Isotope values, latitude, longitude and more data from 357 breeding Wilson's warblers}
\format{
A tibble with 357 rows and 15 variables. The relevant variables for
analyses here are:
\describe{
  \item{ID}{unique identifier for each bird}
  \item{Isotope.Value}{hydrogen isotope ratios measured in the bird's feather}
  \item{lat}{latitude of the bird's breeding/sampling location}
  \item{long}{latitude of the bird's breeding/sampling location}
}
}
\source{
Kristen Ruegg, Jeff Kelly, Thomas Smith
}
\usage{
breeding_wiwa_isotopes
}
\description{
A data frame containing hydrogen isotope values, lat, long, and IDs and some other
columns of data for birds sampled on the breeding grounds. Notice that the latitude
column is named "lat" and the longitude column is named "long".  Those names are both
all lowercase.  That is the way we roll here.  Make sure that you use "lat" and "long" instead
of "Lat" and "Long".
}
\keyword{datasets}
