% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine-types.R
\name{comboize_and_fortify}
\alias{comboize_and_fortify}
\title{prepare fortified output for multipanel plot}
\usage{
comboize_and_fortify(mgen, miso, mhab, true_lat = NA, true_long = NA,
  gen_beta_levels = 1, iso_beta_levels = c(1), hab_beta_levels = c(1))
}
\arguments{
\item{true_lat}{the true latitude where the bird was sampled.}

\item{true_long}{the true longitude where the bird was sampled.}

\item{Mgen}{genetics posterior raster}

\item{Miso}{isotope posterior raster}

\item{Mhab}{habitat suitability raster}
}
\description{
This takes Mgen, Miso, and Mhab for a single bird
and, if available, the true breeding location.  Then it
computes the combo-ized raster at all the requested levels
of the exponents, and creates a fortified data frame of the
results suitable for plotting in ggplot
}

