% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaiah.R
\docType{package}
\name{gaiah}
\alias{gaiah}
\title{gaiah: Genetic and Isotopic Assignment Accounting for Habitat Suitability}
\description{
Tools for using genetic markers, stable isotope data, and habitat
suitability data to calculate posterior probabilities of breeding origin of
migrating birds.
}
\details{
There is not a tutorial within the package, currently.  The best place to find an
example of how to use the functions in this package is at the GitHub repository:
\url{https://github.com/eriqande/gaiah-wiwa}. Go ahead and read the README there.
It should provide you with everything you need to get up and running with
the gaiah package.

Finally, note that the development version of gaiah is available at
\url{https://github.com/eriqande/gaiah}.
}
