% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{migrant_wiwa_isotopes}
\alias{migrant_wiwa_isotopes}
\title{Isotope values and meta data for 688 migrating Wilson's Warblers}
\format{
A tibble with 688 rows and 14 variables. The relevant variables for
analyses here are:
\describe{
  \item{ID}{unique identifier for each bird}
  \item{Isotope.Value}{hydrogen isotope ratios measured in the bird's feather}
}
}
\source{
Kristina Paxton
}
\usage{
migrant_wiwa_isotopes
}
\description{
A data frame containing hydrogen isotope values, lat, long, and IDs and some other
columns of data for birds sampled during migration from Arizona.  604 of the individuals
in this data set also have values in \code{\link{migrant_wiwa_genetic_posteriors}}.
}
\keyword{datasets}
